+++ MySession/src/main/java/configs/Base.java
@@ -52,8 +52,7 @@ public class Base {
 
 	@BeforeMethod(alwaysRun = true)
 	public void methodName(Method method) {
-		ExtentManager.parentTest = ExtentManager.extent.createTest(method.getName());
-		Log.startTestCase(method.getName());
+		
 	}
 
 	@BeforeSuite(alwaysRun = true)
@@ -107,18 +106,5 @@ public class Base {
 		ExtentManager.endReport();
 	}
 
-	public String screenShot(WebDriver driver, String filename) {
-		String dateName = new SimpleDateFormat("yyyyMMddhhmmss").format(new Date());
-		TakesScreenshot takesScreenshot = (TakesScreenshot) driver;
-		File source = takesScreenshot.getScreenshotAs(OutputType.FILE);
-		String destination = System.getProperty("user.dir") + "\\ScreenShots\\" + filename + "_" + dateName + ".png";
 
-		try {
-			FileUtils.copyFile(source, new File(destination));
-		} catch (Exception e) {
-			e.getMessage();
-		}
-
-		return destination;
-	}
 }
