+++ MySession/src/main/java/utilities/Listener.java
@@ -1,12 +1,16 @@
 package utilities;
 
 import java.io.File;
-import java.io.IOException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
 import org.apache.commons.io.FileUtils;
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
+import org.openqa.selenium.WebDriver;
 import org.testng.ITestContext;
 import org.testng.ITestListener;
 import org.testng.ITestResult;
-import com.aventstack.extentreports.MediaEntityBuilder;
 import com.aventstack.extentreports.Status;
 import com.aventstack.extentreports.markuputils.ExtentColor;
 import com.aventstack.extentreports.markuputils.MarkupHelper;
@@ -20,12 +24,14 @@ public class Listener extends ExtentManager implements ITestListener {
 	//Action action= new Action();
 	
 	public void onTestStart(ITestResult result) {
-		//parentTest = extent.createTest(result.getName());
+		parentTest = extent.createTest(result.getName());
+		Log.startTestCase(result.getName());
 	}
 
 	public void onTestSuccess(ITestResult result) {
 		if (result.getStatus() == ITestResult.SUCCESS) {
 			parentTest.log(Status.PASS, "Pass Test case is: " + result.getName());
+			Log.endTestCase(result.getName());
 		}
 	}
 
@@ -36,8 +42,9 @@ public class Listener extends ExtentManager implements ITestListener {
 			parentTest.log(Status.FAIL,
 					MarkupHelper.createLabel(result.getThrowable() + " - Test Case Failed", ExtentColor.RED));
 			String imgPath = screenShot(Base.getDriver(), result.getName());
-
-			parentTest.fail("ScreenShot is Attached", MediaEntityBuilder.createScreenCaptureFromPath(imgPath).build());
+			parentTest.addScreenCaptureFromPath(imgPath);
+			Log.error("Test FAILED");
+			 //MediaEntityBuilder.createScreenCaptureFromPath(imgPath).build();
 		}
 	}
 
@@ -57,10 +64,23 @@ public class Listener extends ExtentManager implements ITestListener {
 	}
 
 	public void onFinish(ITestContext context) {
-		
+		Log.info("TEST finished");
 	}
 
+	public String screenShot(WebDriver driver, String filename) {
+		String dateName = new SimpleDateFormat("yyyyMMddhhmmss").format(new Date());
+		TakesScreenshot takesScreenshot = (TakesScreenshot) driver;
+		File source = takesScreenshot.getScreenshotAs(OutputType.FILE);
+		String destination = System.getProperty("user.dir") + "\\ScreenShots\\" + filename + "_" + dateName + ".png";
 
+		try {
+			FileUtils.copyFile(source, new File(destination));
+		} catch (Exception e) {
+			e.getMessage();
+		}
+
+		return destination;
+	}
 		
 	}
