+++ MySession/test-output/old/Default suite/groups.html
@@ -1,9 +1,4 @@
 <h2>Groups used for this test run</h2><table border="1">
-<<<<<<< HEAD
-<tr> <td align="center"><b>Group name</b></td><td align="center"><b>Methods</b></td></tr><tr><td>login</td><td>LoginTest.loginTest()[pri:0, instance:jarvis.tests.LoginTest@e15b7e8]<br/></td></tr>
-<tr><td>regression</td><td>LoginTest.loginTest()[pri:0, instance:jarvis.tests.LoginTest@e15b7e8]<br/></td></tr>
-=======
-<tr> <td align="center"><b>Group name</b></td><td align="center"><b>Methods</b></td></tr><tr><td>login</td><td>LoginTest.loginTest()[pri:0, instance:jarvis.tests.LoginTest@6fd83fc1]<br/></td></tr>
-<tr><td>regression</td><td>LoginTest.loginTest()[pri:0, instance:jarvis.tests.LoginTest@6fd83fc1]<br/></td></tr>
->>>>>>> branch 'dev' of https://github.com/arslanmomin/Expleo_Jarvis.git
+<tr> <td align="center"><b>Group name</b></td><td align="center"><b>Methods</b></td></tr><tr><td>login</td><td>LoginTest.loginTest()[pri:0, instance:jarvis.tests.LoginTest@1b2abca6]<br/></td></tr>
+<tr><td>regression</td><td>LoginTest.loginTest()[pri:0, instance:jarvis.tests.LoginTest@1b2abca6]<br/></td></tr>
 </table>
