+++ MySession/test-output/old/Default suite/methods-alphabetical.html
@@ -1,27 +1,14 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Default suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<<<<<<< HEAD
-<tr bgcolor="e5eea7">  <td>22/05/11 16:24:33</td>   <td>0</td> <td title="&lt;&lt;Base.afterClass()[pri:0, instance:jarvis.tests.LoginTest@e15b7e8]">&lt;&lt;afterClass</td> 
-<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@20853837</td>   <td></td> </tr>
-<tr bgcolor="e5eea7">  <td>22/05/11 16:24:20</td>   <td>-13417</td> <td title="&gt;&gt;Base.beforeSuite()[pri:0, instance:jarvis.tests.LoginTest@e15b7e8]">&gt;&gt;beforeSuite</td> 
-<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@20853837</td>   <td></td> </tr>
-<tr bgcolor="7e99d8">  <td>22/05/11 16:24:30</td>   <td>-2762</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="LoginTest.loginTest()[pri:0, instance:jarvis.tests.LoginTest@e15b7e8]">loginTest</td> 
-  <td>main@20853837</td>   <td></td> </tr>
-<tr bgcolor="e5eea7">  <td>22/05/11 16:24:30</td>   <td>-2771</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;Base.methodName(java.lang.reflect.Method)[pri:0, instance:jarvis.tests.LoginTest@e15b7e8]">&gt;&gt;methodName</td> 
-<td>&nbsp;</td>  <td>main@20853837</td>   <td></td> </tr>
-<tr bgcolor="e5eea7">  <td>22/05/11 16:24:30</td>   <td>-3035</td> <td>&nbsp;</td><td title="&gt;&gt;Base.report()[pri:0, instance:jarvis.tests.LoginTest@e15b7e8]">&gt;&gt;report</td> 
-<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@20853837</td>   <td></td> </tr>
-=======
-<tr bgcolor="e5eea7">  <td>22/05/11 10:42:31</td>   <td>0</td> <td title="&lt;&lt;Base.afterClass()[pri:0, instance:jarvis.tests.LoginTest@6fd83fc1]">&lt;&lt;afterClass</td> 
-<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1101048445</td>   <td></td> </tr>
-<tr bgcolor="e5eea7">  <td>22/05/11 10:42:19</td>   <td>-11592</td> <td title="&gt;&gt;Base.beforeSuite()[pri:0, instance:jarvis.tests.LoginTest@6fd83fc1]">&gt;&gt;beforeSuite</td> 
-<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1101048445</td>   <td></td> </tr>
-<tr bgcolor="e5eea7">  <td>22/05/11 10:42:27</td>   <td>-3866</td> <td>&nbsp;</td><td title="&gt;&gt;Base.generateReport()[pri:0, instance:jarvis.tests.LoginTest@6fd83fc1]">&gt;&gt;generateReport</td> 
-<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1101048445</td>   <td></td> </tr>
-<tr bgcolor="7e99d8">  <td>22/05/11 10:42:27</td>   <td>-3231</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="LoginTest.loginTest()[pri:0, instance:jarvis.tests.LoginTest@6fd83fc1]">loginTest</td> 
-  <td>main@1101048445</td>   <td></td> </tr>
-<tr bgcolor="e5eea7">  <td>22/05/11 10:42:27</td>   <td>-3244</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;Base.methodName(java.lang.reflect.Method)[pri:0, instance:jarvis.tests.LoginTest@6fd83fc1]">&gt;&gt;methodName</td> 
-<td>&nbsp;</td>  <td>main@1101048445</td>   <td></td> </tr>
->>>>>>> branch 'dev' of https://github.com/arslanmomin/Expleo_Jarvis.git
+<tr bgcolor="e5eea7">  <td>22/05/11 17:50:04</td>   <td>0</td> <td title="&lt;&lt;Base.afterClass()[pri:0, instance:jarvis.tests.LoginTest@1b2abca6]">&lt;&lt;afterClass</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@443384617</td>   <td></td> </tr>
+<tr bgcolor="e5eea7">  <td>22/05/11 17:49:43</td>   <td>-21365</td> <td title="&gt;&gt;Base.beforeSuite()[pri:0, instance:jarvis.tests.LoginTest@1b2abca6]">&gt;&gt;beforeSuite</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@443384617</td>   <td></td> </tr>
+<tr bgcolor="7e99d8">  <td>22/05/11 17:49:52</td>   <td>-12049</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="LoginTest.loginTest()[pri:0, instance:jarvis.tests.LoginTest@1b2abca6]">loginTest</td> 
+  <td>main@443384617</td>   <td></td> </tr>
+<tr bgcolor="e5eea7">  <td>22/05/11 17:49:52</td>   <td>-12051</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;Base.methodName(java.lang.reflect.Method)[pri:0, instance:jarvis.tests.LoginTest@1b2abca6]">&gt;&gt;methodName</td> 
+<td>&nbsp;</td>  <td>main@443384617</td>   <td></td> </tr>
+<tr bgcolor="e5eea7">  <td>22/05/11 17:49:52</td>   <td>-12624</td> <td>&nbsp;</td><td title="&gt;&gt;Base.report()[pri:0, instance:jarvis.tests.LoginTest@1b2abca6]">&gt;&gt;report</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@443384617</td>   <td></td> </tr>
 </table>
