+++ MySession/test-output/old/Default suite/toc.html
@@ -20,9 +20,9 @@
 <td><a target='mainFrame' href='reporter-output.html'>reporter output</a></td>
 <td><a target='mainFrame' href='testng.xml.html'>testng.xml</a></td>
 </tr></table>
-<table width='100%' class='test-passed'>
+<table width='100%' class='test-failed'>
 <tr><td>
-<table style='width: 100%'><tr><td valign='top'>Default test (1/0/0)</td><td valign='top' align='right'>
+<table style='width: 100%'><tr><td valign='top'>Default test (0/1/0)</td><td valign='top' align='right'>
   <a href='Default test.html' target='mainFrame'>Results</a>
 </td></tr></table>
 </td></tr><p/>
